@@ -1,4 +1,3 @@
-import { useRouter } from "next/router";
 import { FC } from "react";
 
 import { createElementSmartLink, createItemSmartLink } from "../../lib/utils/smartLinkUtils";
@@ -11,12 +10,12 @@ type ElementCodename = Readonly<{
   elementCodename: string;
 }>;
 
-type Props = ItemId | ElementCodename;
+type Props = (ItemId | ElementCodename)  & {
+  isPreview?: boolean //TODO: when converted to app router change this to non-undefined type
+};
 
 export const StandaloneSmartLinkButton: FC<Props> = props => {
-  const isPreview = useRouter().isPreview;
-
-  if (!isPreview) {
+  if (!props.isPreview) {
     return null;
   }
 
